% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_background}
\alias{calculate_background}
\title{Calculate background area}
\usage{
calculate_background(data, area, criteria = name \%in\% c("background",
  "Background"), update_category = "background", quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{area}{the area column}

\item{criteria}{the expression used to find the background analyses}

\item{update_category}{value to update the found records category to
(pass in NULL if no update of category is desired)}
}
\value{
adds the parameter column p.bgrd
}
\description{
Calculates the bacterial background area from the direct measurement of the
background peak. For setting this parameter manually, please use
\code{\link{set_background}}.
}
\seealso{
\code{\link{change_category}}
}

