% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_concentrations}
\alias{calculate_concentrations}
\title{Calculate the concentrations of the source samples}
\usage{
calculate_concentrations(data, area, volume, dilution = 1,
  conc_pattern = "(\\\\d+)uM", standards = category \%in\% c("USGS-34",
  "IAEA-NO3"), quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{area}{the area column}

\item{volume}{the volume column}

\item{dilution}{the dilution column (if any)}

\item{conc_pattern}{the regular expression pattern to capture the concentration
from the standards 'name' colum}

\item{standards}{selection criterion for the standards to consider}
}
\value{
introduces the columns 'conc', 'amount' and the parameters 'p.yield' and 'p.run_size'
}
\description{
Calculate the concentrations of the source samples
}

