% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_oxidation_blank}
\alias{calculate_oxidation_blank}
\title{Calculate oxidation blank area/volume}
\usage{
calculate_oxidation_blank(data, area, volume, criteria = grepl("POR Blank",
  name), update_category = "oxidation blank", quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{area}{the area column}

\item{volume}{the volume column}

\item{criteria}{the expression used to find the analyses}

\item{update_category}{value to update the found records category to
(pass in NULL if no update of category is desired)}
}
\value{
adds the parameter column p.oblank
}
\description{
Calculates the oxidation blank area to volume ratio from the direct
measurements of the blank peak. For setting this parameter manually, please use
\code{\link{set_oxidation_blank}}.
}
\seealso{
\code{\link{change_category}}
}
