% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calibrate_d15}
\alias{calibrate_d15}
\title{Calibrate d15 values with the given standards}
\usage{
calibrate_d15(data, d15, standards = c(`USGS-34` = -1.8, `IAEA-NO3` = 4.7),
  quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{d15}{the d15 column}

\item{standards}{a set of isotope standards
Note: they are matched to the data by "category" (not by name)}
}
\value{
introduces the column d15.cal and parameters p.d15_m and p.d15_b + p.15_stds
}
\description{
Calibrate d15 values with the given standards
}
\note{
implement single point correction
}

