% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calibrate_d15_org}
\alias{calibrate_d15_org}
\title{Calibrate organic d15 values with the given standards}
\usage{
calibrate_d15_org(data, d15, volume = volume, standards = c(`USGS-40` =
  -4.5, `USGS-41` = 47.6), infer_NO3_blank = FALSE, ref_gas_d15 = 0,
  quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{d15}{the d15 column (can be a column already calibrated for denitrifier method)}

\item{standards}{a set of isotope standards
Note: they are matched to the data by "category" (not by name)}

\item{infer_NO3_blank}{whether to infer the NO3 blank from this correction in addition to the organic blank (if using a single step correction)}

\item{ref_gas_d15}{the reference gas isotopic composition (only used if infer_NO3_blank = TRUE, to constrain culture blank d15)}
}
\value{
introduces the column d15.ocal and parameters p.15o_stds,
   p.d15o_m_true, p.d15o_m_vol, p.d15o_m_true:vol and p.d15o_b
   as well as information on the organic blank p.No_blk_conc and p.No_blk_d15 and nitrate blank (only if type is 1-step!) p.N_blk_amt and p.N_blk_d15
}
\description{
Calibrate organic d15 values with the given standards
}
\note{
implement single point correction
}

