% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{change_category}
\alias{change_category}
\title{Change the category of specific analyses}
\usage{
change_category(data, criteria, value, quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{criteria}{an expression to identify records whose category should be change to the new value}

\item{value}{the new category value for the records matching the criteria}
}
\description{
This can be used to assign new categories or change existing ones based on an expression that is evaluated in the context of the data frame.
}

