% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{evaluate_drift}
\alias{evaluate_drift}
\title{Evaluate drift}
\usage{
evaluate_drift(data, d15, d18, group = name, correct = FALSE,
  correct_with = category \%in\% c("USGS-34", "IAEA-NO3", "N2O"),
  plot = TRUE, quiet = FALSE, method = "lm", ...)
}
\arguments{
\item{data}{the data frame, must be from a single run because this should really be evaluated
for individual runs}

\item{group}{expression for what to group by in order to normalize properly
this is also used for color coding the data points in the generated plot (if \code{plot = TRUE})}

\item{correct}{whether to correct for drift or not (default is FALSE)}

\item{correct_with}{expression for what all to include in the correction}

\item{plot}{whether to output the drift plots [default is FALSE]}

\item{method}{which method to use for drift correction, common approaches
are linear models (lm) and local polynomial regression fitting (loess)}

\item{...}{additional parameters for the fitting method}
}
\value{
introduces the new columns d15.cor and d18.cor + parameter column p.drift
}
\description{
This function evaluates drift and can correct for it (use paramter \code{correct = TRUE}).
}

