% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_parameter_table}
\alias{generate_parameter_table}
\title{generate a summary of the parameters (lists all unique rows so make sure to group appropriately!)}
\usage{
generate_parameter_table(data)
}
\arguments{
\item{the}{data frame, can already have a group_by if so desired}
}
\description{
generate a summary of the parameters (lists all unique rows so make sure to group appropriately!)
}
