% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_calib_coef}
\alias{get_calib_coef}
\title{method making it easier to retrieve regression coefficients by name}
\usage{
get_calib_coef(key, model, name)
}
\description{
method making it easier to retrieve regression coefficients by name
}

