% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_isodat_data_tables}
\alias{get_isodat_data_tables}
\title{Combine isodat files' data tables}
\usage{
get_isodat_data_tables(isodat_files, quiet = FALSE)
}
\arguments{
\item{isodat_files}{isodat objects}

\item{quiet}{whether the function should output information messages or be quiet (default is to output)}
}
\description{
Combine isodat files' data tables
}

