% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isorunN2O-package}
\alias{isorunN2O}
\alias{isorunN2O-package}
\title{isorunN2O: Process IRMS runs of N2O analyses.}
\description{
This package facilitates the reading, processing and visualization
    of N2O isotope data including drift correction, 17O correction and standards
    calibration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebkopf/isorunN2O}
  \item Report bugs at \url{https://github.com/sebkopf/isorunN2O/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
