% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isorunN2O}
\alias{isorunN2O}
\alias{isorunN2O-package}
\title{Process IRMS runs of N2O analyses}
\description{
This package facilitates the reading, processing and visualization of N2O isotope data including drift correction, 17O correction and standards calibration.

To get started, take alook at the tutorial R markdown file.
}
\author{
Sebastian Kopf
}

