% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{last_plot}
\alias{last_plot}
\title{make sure the last plot uses ggplot instead of plotly implementation}
\usage{
last_plot()
}
\description{
make sure the last plot uses ggplot instead of plotly implementation
}
\note{
not sure this is the best way to deal with this, temporary fix while plotly causes trouble
}

