% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_itext}
\alias{make_itext}
\title{Generate more elaborate interactive hover text labels}
\usage{
make_itext(...)
}
\arguments{
\item{...}{expression what to use for the interactive text
for example: make_itext(name, d45 = round(d45, 2), ...)}
}
\description{
Generate more elaborate interactive hover text labels
}

