% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{run_d15_calibration}
\alias{run_d15_calibration}
\title{runs the d15 calibration internally
called from calibrate_d15 and calibrate_d15_org}
\usage{
run_d15_calibration(data, standards, organic, infer_ref_gas = TRUE,
  infer_bgrd = TRUE, infer_oblank = TRUE, quiet = FALSE)
}
\description{
runs the d15 calibration internally
called from calibrate_d15 and calibrate_d15_org
}

