% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viewer.R
\name{run_data_viewer}
\alias{run_data_viewer}
\title{Run the N2O data viewer}
\usage{
run_data_viewer(base_dir = ".", ...)
}
\arguments{
\item{base_dir}{the base data directory where the data viewer should start}

\item{...}{passed on to the \link{shiny::runApp} call and could include paramters such as host or port}
}
\description{
To run this function stand-alone as a script from any folder on your system, copy the
run (Unix systems) or run.bat (Windows) systems from the GitHub repository
(\link{http://www.github.com/sebkopf/isorunN2O/inst/shiny-apps/data_viewer/}
and modify depending on your R executable.
}

