% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viewer.R
\name{run_data_viewer_dev}
\alias{run_data_viewer_dev}
\title{for development of data viewer
not exported to namespace but helpful for active development of the shiny app}
\usage{
run_data_viewer_dev(base_dir = ".", app_dir = file.path("inst",
  "shiny-apps", "data_viewer"), ...)
}
\description{
for development of data viewer
not exported to namespace but helpful for active development of the shiny app
}
