% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{select_columns}
\alias{select_columns}
\title{Select relevant data columns}
\usage{
select_columns(data, ..., quiet = FALSE)
}
\arguments{
\item{data}{data frame}

\item{quiet}{whether the function should output information messages or be quiet (default is to output)}
}
\description{
This is just a convenience wrapper for \code{\link{select}} but one that provides helpful information about the selected and discarded columns.
}
