% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{set_background}
\alias{set_background}
\title{Set the bacterial background}
\usage{
set_background(data, value, quiet = FALSE)
}
\arguments{
\item{data}{the data set}

\item{value}{the value of the bacterial blank area, can be an expression}
}
\value{
adds the parameter column p.bgrd
}
\description{
This can be calculated automatically based on identifying which analyses are
measurements of the background alone (using \code{\link{calculate_background}}).
This function is for setting this parameter manually in cases where the background
is estimated or the peak was too small to be automatically recorded.
}
\note{
uses non-standard evaluation
}

