% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{set_oxidation_blank}
\alias{set_oxidation_blank}
\title{Set the oxdiation blank arae/volume}
\usage{
set_oxidation_blank(data, value, quiet = FALSE)
}
\arguments{
\item{data}{the data set}

\item{value}{the numeric value of the oxidation blank area/volume ratio}
}
\value{
adds the parameter column p.oblank
}
\description{
This can be calculated automatically based on identifying which analyses are
measurements of the oxidation blank alone (using \code{\link{calculate_oxidation_blank}}).
This function is for setting this parameter manually in cases where the background
is estimated or the peak was too small to be automatically recorded.
}

