% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{set_run_size}
\alias{set_run_size}
\title{Set the run size}
\usage{
set_run_size(data, amount, quiet = FALSE)
}
\arguments{
\item{data}{the data set}

\item{amount}{the run size, should be in the same units as standard concentration * volume injection, i.e. if concentrations are in microM and volumes in mL, the run size should be in nmol}
}
\value{
adds/overwrites the column 'p.run_size'
}
\description{
Not that this is calculated automatically when using calculate_concentrations,
only use this function to set the run size manually or overwrite automatic
calculations from calculate_concentrations.
}

