% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calibrate_d18}
\alias{calibrate_d18}
\title{Calibrate d18 data with the given standards.}
\usage{
calibrate_d18(data, d18, amount = amount, volume = volume, cell_volume,
  standards = c(`USGS-34` = -27.93, `IAEA-NO3` = 25.61), quiet = FALSE)
}
\arguments{
\item{data}{(can be a grouped_by data set)}

\item{d18}{the d18 column}

\item{amount}{the amount column}

\item{volume}{the volume column}

\item{cell_volume}{the volume of cells in the vial (has to be same units as volume!)}

\item{standards}{a set of isotope standards
Note: they are matched to the data by "category" (not by name)}
}
\value{
introduces the column d18.cal and parameters p.d18_m_true, p.d18_m_conc, p.d18_m_true:conc and p.d18_b
}
\description{
Uses a multivariate linear regression that takes the effective concentration
in the sample vials into consideration (as well as covariance between standards
and effective concentration).
}

