% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_interactive}
\alias{make_interactive}
\title{Makes a ggplot interactive using ggplotly.}
\usage{
make_interactive(p = ggplot2:::last_plot(), theme = theme_grey(),
  plus = NULL)
}
\arguments{
\item{p}{the ggplot to make interactive, by default the last plot}

\item{theme}{the theme to apply, theme_grey is default (theme_bw is hard to read)}

\item{plus}{additional ggplot items to add to the plot}
}
\description{
Note that in order for these plots to render properly in knitted RMarkdown documents, you have to actually load the plotly library in your code. Not doing it in this function itself because it's generally considered bad practice.
}

