% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{select_N2O_peak}
\alias{select_N2O_peak}
\title{Identify the N2O peak}
\usage{
select_N2O_peak(data_table, peak_rt, quiet = FALSE)
}
\arguments{
\item{data_table}{aggregated isodat data table(s)}

\item{peak_rt}{peak retention time [in s] - either a single value, which will find peaks that start before and end after this retention time; or a vector with two values: c(from, to) which will find peaks that overlap with this interval}

\item{quiet}{whether the function should output information messages or be quiet (default is to output)}
}
\description{
Identify the N2O peak
}
